挪动 loadSetting 到 utils
